+++ src/Console.cpp
@@ -19,6 +19,54 @@ bool Console::focusInput = false;
 unsigned long Console::lastLogLength = 0;
 std::vector<std::string> Console::lastCommands;
 long Console::lastCommandIndex = -1;
+std::string Console::bufferedCommand;
+
+void Console::callback(ImGuiTextEditCallbackData* data) {
+    bool update = false;
+    std::string completion;
+
+    switch (data->EventKey) {
+        case ImGuiKey_Tab:
+            completion = Command::autoComplete(data->Buf);
+            if (completion.size() > 0) {
+                data->DeleteChars(0, data->BufSize);
+                data->InsertChars(0, completion.c_str());
+                data->CursorPos = strlen(data->Buf);
+            }
+            break;
+
+        case ImGuiKey_UpArrow:
+            update = true;
+            if (lastCommandIndex < (long)(lastCommands.size() - 1)) {
+                lastCommandIndex++;
+                if (lastCommandIndex == 0)
+                    bufferedCommand = data->Buf;
+            }
+            break;
+
+        case ImGuiKey_DownArrow:
+            update = true;
+            if (lastCommandIndex > -1)
+                lastCommandIndex--;
+            break;
+    }
+
+    if (update) {
+        if ((lastCommandIndex >= 0) && (lastCommandIndex < lastCommands.size())) {
+            if (lastCommands.at(lastCommands.size() - 1 - lastCommandIndex) != buffer) {
+                data->DeleteChars(0, data->BufSize);
+                data->InsertChars(0, lastCommands.at(lastCommands.size() - 1 - lastCommandIndex).c_str());
+            }
+        } else {
+            data->DeleteChars(0, data->BufSize);
+            if (bufferedCommand.size() > 0) {
+                data->InsertChars(0, bufferedCommand.c_str());
+            }
+        }
+
+        data->CursorPos = strlen(data->Buf);
+    }
+}
 
 void Console::display() {
     if (ImGui::Begin("Console", NULL, ImVec2(600, 400), -1.0f)) {
@@ -37,61 +85,32 @@ void Console::display() {
         }
         ImGui::EndChild();
 
-        if (focusInput) {
-            ImGui::SetKeyboardFocusHere();
-            focusInput = false;
-        }
-
-        if (ImGui::GetWindowIsFocused()) {
-            ImGuiIO& io = ImGui::GetIO();
-            static bool hold = false;
-            bool update = false;
-
-            if ((!hold) && (io.KeysDown[upKey] || io.KeysDown[downKey]))
-                update = hold = true;
-
-            if (hold && (!(io.KeysDown[upKey] || io.KeysDown[downKey])))
-                hold = false;
-
-            if (update) {
-                if (io.KeysDown[upKey]) {
-                    if (lastCommandIndex < static_cast<long>(lastCommands.size() - 1))
-                        lastCommandIndex++;
-                }
-
-                if (io.KeysDown[downKey]) {
-                    if (lastCommandIndex > -1)
-                        lastCommandIndex--;
-                }
-
-                if ((lastCommandIndex >= 0) && (lastCommandIndex < lastCommands.size())) {
-                    strncpy(buffer,
-                            lastCommands.at(lastCommands.size() - 1 - lastCommandIndex).c_str(),
-                            bufferLength);
-                    buffer[bufferLength] = '\0';
-                } else {
-                    buffer[0] = '\0';
-                }
-
-                update = false;
-            }
-        }
-
         if (ImGui::InputText("Command", buffer, bufferLength,
-                             ImGuiInputTextFlags_AutoSelectAll | ImGuiInputTextFlags_EnterReturnsTrue)) {
+                             ImGuiInputTextFlags_EnterReturnsTrue
+                              | ImGuiInputTextFlags_CallbackCompletion
+                              | ImGuiInputTextFlags_CallbackHistory,
+                             &Console::callback)) {
             getLog() << "> " << buffer << Log::endl;
-            int error = Command::command(buffer);
-            if (error != 0) {
-                getLog() << "Error code: " << error << Log::endl;
+            if (strlen(buffer) > 0) {
+                int error = Command::command(buffer);
+                if (error != 0) {
+                    getLog() << "Error code: " << error << Log::endl;
+                }
+
+                if ((lastCommands.size() == 0) || (lastCommands[lastCommands.size() - 1] != buffer))
+                    lastCommands.push_back(std::string(buffer));
             }
 
-            if ((lastCommands.size() == 0) || (lastCommands[lastCommands.size() - 1] != buffer))
-                lastCommands.push_back(std::string(buffer));
             lastCommandIndex = -1;
-
             buffer[0] = '\0';
             scrollToBottom = true;
             focusInput = true;
+            bufferedCommand.clear();
+        }
+
+        if (ImGui::IsItemHovered() || focusInput) {
+            ImGui::SetKeyboardFocusHere(-1);
+            focusInput = false;
         }
     }
     ImGui::End();